Add test for encrypt/decrypt message, read/write multiple chunk message, make authentication challenge code, make signed authentication message(proto), read/check signed authentication message.
Refactor some test to use .expect instead of assert on result.

Add identifier for peer and option for an address.

Split the make authentication challenge code out of read/write authentication.

Fix the bug where I try to decode the message based on shared secret instead of the authentication challenge code.

Check if the address of the peer match the verification key used if provided.